ntools_iaea

Function fetch_all

pub fn fetch_all(rad_type: RadType) -> HashMap<String, RecordSet>

Fetch all nuclides direct from IAEA

This will make ~3500 requests

This will request and process all available nuclides from the chart of nuclides. This is fully parallelised, but will still take a few minutes to complete.

It is strongly recommended that fetch_nuclide() or fetch_nuclides() are used instead for more targeted fetch requests.

Failure to collect a list of all known nuclides will panic and exit the program because this is crucial missing information.

Specify the type of decay data with the appropriate RadType to deserialise the entire chart of nuclides.

// Get all records for all nuclides available in the chart of nuclides
let decay_data: HashMap<String, RecordSet> = fetch_all(RadType::Gamma);

The returned hashmap is a dictionary of key value pairs where:

key : Nuclide name, e.g. “Co60”

value : List of all matching Records