ntools_iaea

Function prefetch_json

pub fn prefetch_json<P: AsRef<Path>>(
    path: P,
    rad_type: RadType,
) -> Result<(), Error>

Generate a JSON file for pre-fetched data

This will use fetch_all() and is therefore parallelised, but may still take up to a few minutes depending on radiation type.

The JSON format is human-readable and simple to parse, but will use much more space on disk and be slower when retrieving records.

// Fetch all chart of nuclide data and store as JSON formatted file
prefetch_json("/path/to/file.json", RadType::Gamma).unwrap();